chore: move types to separate file
